eth_ocm: fix the last tx byte corruption
Jake had fixed the bug.

The "eth_avalon_txdma.v" module has two functions. One is to transfer
data from memory to the MAC core for transmit. The other is to report
transmit status back to the processor via the descriptor. The DMA
controller supports unaligned memory transfers and transfers of any
length.

The memory interface is 32 bits (4 bytes). So when transferring data to
the MAC core, the DMA controller must determine which of the 4 bytes
actually contain valid transmit data. This is dependent on the starting
read address and the transmit length. For the vast majority of the data,
all 4 bytes are valid. The two special cases occur on the first and last
reads from memory. The first read from memory may only contain partial
valid data based on the starting read address. The last read from memory
may only contain partial data based on the starting read address and the
transmit length.

The problem obviously was occurring on the last data read. When I start
a data transfer, I store the transmit length from the descriptor.
However, when I finally got to the last data read and passed the number
of valid bytes to the transmit section of the core, I was looking at the
descriptor input rather than the stored descriptor to determine the
length. The descriptor input at that point has about a 50% chance of
being incorrect which means the MAC was probably corrupting packets at a
rate of 50%. This is why it was difficult to reproduce.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>