eth_ocm: New version (8.03) of the core.
Oyvind Harboe began implementing the MAC with eCOS. In the process he
uncovered several issues with the MAC implementation:

1 - Aborted packets were not being handled properly by the RX DMA
controller. As a result, even aborted packets were being passed to the
software layer for processing.

2 - A transmit problem was discovered with memory interfaces that have
very long read latencies. If a small packet was being transmitted, it
was possible for the DMA controller to issue all of the reads for the
packet to the Avalon bus and then begin transmission before any of the
read data actually came back. The result was that the packet would not
be transmitted at all.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>